fix: docs and sdam_viz move
